move gymanx compat to builtin
